+++ Backend/.env.example
@@ -1,1 +1,9 @@
-DATABASE_URL="sqlserver://SIDDHESH:1443;database=task;user=sa;password=root;encrypt=true;trustServerCertificate=true"
+DB_USER=sa
+DB_HOST=SIDDHESH:1443
+DB_PASSWORD=root
+DB_NAME=task
+DB_DIALECT=sqlserver
+PORT=5000
+
+
+DATABASE_CONNECTION_URL="sqlserver://${DB_HOST};database=${DB_NAME};user=${DB_USER};password=${DB_PASSWORD};encrypt=true;trustServerCertificate=true"
