+++ Backend/package.json
@@ -19,7 +19,6 @@
     "mysql": "^2.18.1",
     "mysql2": "^3.11.0",
     "path": "^0.12.7",
-    "sequelize": "^6.37.3",
     "@prisma/client": "^5.19.1"
   },
   "devDependencies": {
