--- Backend/src/constants/AUTH/registerConstants.ts
@@ -1,8 +1,0 @@
-export const REGISTER_MESSAGES = {
-    CREDENTIALS_NEEDED: "Credentials Needed",
-    USER_ALREADY_EXISTS: "User already exists",
-    ERROR_REGISTERING_USER: "Error registering user",
-};
-
-
-export const JWT_SECRET = process.env.JWT_SECRET || "defaultSecret";
