--- Backend/src/enums/index.ts
@@ -1,4 +1,0 @@
-export enum DbModelsEnum{
-    USERS= "user",
-    TODOS= "todo",
-}
