+++ Backend/src/server.ts
@@ -1,9 +1,9 @@
-import express, { Request, Response, NextFunction } from "express";
+import express from "express";
 import cors from "cors";
 import dotenv from "dotenv";
-
+import db from "./db/connect-db";
 import todosRouter from "./router/todo.router";
-import userRouter from "./router/user.router";
+import userRouter from "./router/user-auth.router";
 
 dotenv.config();
 
@@ -13,9 +13,12 @@ app.use(express.json());
 app.use(userRouter);
 app.use(todosRouter);
 
-
-const PORT = process.env.PORT;
-
-app.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}`);
+db.then(() => {
+  const PORT = process.env.PORT || 5000;
+  app.listen(PORT, () => {
+    console.log(`Server is running on port ${PORT}`);
+  });
+}).catch((err) => {
+  console.error("Failed to start the server due to database connection issues", err);
+  process.exit(1); 
 });
