--- Frontend/.vite/deps_temp_68ab4435/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
